Guess `ExchangeType` if it is not set in `RequestOptions` (#5517)
Motivation:

`ExchangeType` is a hint for performance optimization. If no
`RequestOptions` is set, it is inferred from the type of `HttpRequest`.
However, a custom `RequestOptions` is set, `ExchangeType` is not
inferred even if the `RequestOptions.exchangeType()` is null.

Modifications:

- Use `ExchangeType.RESPONSE_STREAMING` if
`RequestOptions.exchangeType()` is null and `HttpRequest` is an instance
of `FixedStreamMessage`.
- If none of the above conditions are met, the method defaults to
returning ExchangeType.BIDI_STREAMING.


Result:

`ExchangeType` is guessed correctly even if `RequestOptions` is set.