Websocket bug: PING to server sends PING Frame instead of PONG Frame (#5533)
Motivation:

There is a bug in the Websocket Server PING/PONG handing. When a PING message is sent to Server, the Server incorrectly sends out another PING message as a reply. This is because the code to create the PONG message has a bug in it. The problem is that when testing the PING/PONG mechanism, the Server incorrectly send a PING as a reply for a PING. And this caused whoever was sending the initial PING to receive an incorrect reply to the PING. This was only happening for the case when the PING message has data.

Modifications:

- In the method: `com.linecorp.armeria.common.websocket.WebSocketFrame#ofPong(byte[])`, there is the bug. I modified the code to return the correct Frame. See the files changed.
- I also created unit test for this case

![image](https://github.com/line/armeria/assets/104100/3aa51cb2-d3d3-4e23-9350-e808b51f1833)


Result:

- Websocker PING that contain data are now correctly handled and a PONG Frame sent out.
- This means that when testing with a [websocket cli tool](https://github.com/vi/websocat), I now get the correct frame.
BEFORE:
```
$ websocat --print-ping-rtts --ping-interval 5 ws://localhost:8080/ws
[WARN  websocat::ws_peer] Received a pong with a strange content from websocket
[WARN  websocat::ws_peer] Received a pong with a strange content from websocket
[WARN  websocat::ws_peer] Received a pong with a strange content from websocket
```

AFTER:
```
$ websocat --print-ping-rtts --ping-interval 5 ws://localhost:8080/ws
RTT 0.003123 s
RTT 0.000542 s
RTT 0.000530 s

```

<!--
Visit this URL to learn more about how to write a pull request description:
https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->