Add support for self config source (#5459)
Motivation:

This pull request adds support for SELF config source types. This
feature allows the control plane to delegate the decision of which
protocol to use to the client.
Reference:
https://github.com/envoyproxy/envoy/blob/bd18d0fa7790a7e5fe1a95f6ae271ca02614e36c/api/envoy/config/core/v3/config_source.proto#L239

Note that this implementation may not have been completed from envoy
side, so we should use this option with caution.
ref: https://github.com/envoyproxy/envoy/issues/13951

Modifications:

- The dependency relationship between `ResourceNode` and `ConfigSource`
has changed. This can be a problem because the configSource specified in
`ResourceNode#ConfigSource` can be different from the actual
`ConfigSource` used for subscribing. Modified so that `ConfigSource` is
computed before creating a `ResourceNode`.
- Renamed `BootstrapApiConfigs` to `ConfigSourceMapper` since it better
represents the functionality of the class.
- Added `parentConfigSource` to the logic of computing a new
configSource.
- Modified so that subscribed `ResourceNode#configSource` is always
non-null. The only case where this is `null` is for static clusters
(bootstrap clusters)


Result:

- SELF type configSource is now supported

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->