Reuse VirtualHostBuilder on the same hostnamePattern (#5418)
Motivation:

I registered each virtual host with the same hostname, the registered
routes cannot be not found. Because `VirtualHostBuilder` creates new
builder even if hostname is the same. However, it does reuse them on the
same port.
So, I think it's better to reuse them on the same hostname.

Modifications:

- `ServerBuilder.virtualHost(hostnamePattern)` and
`ServerBuilder.virtualHost(defaultHostname, hostnamePattern)` checks
whether there is `builders` with the same hostname. If there is, returns
this builder.

Result:

- Reuse `VirtualHostBuilder` on the same hostname. 
- Even if you register each one on a virtual host with the same
hostname, they will be merged.

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->

---------

Co-authored-by: jrhee17 <guins_j@guins.org>
Co-authored-by: Ikhun Um <ih.pert@gmail.com>
Co-authored-by: minwoox <songmw725@gmail.com>