Fix a possible deadlock when `OAuth2Client` is used with `RetryingClient`
Motivation:

When `OAuth2Client` fails to acquire a granted access token, it directly
returns a failed response. As the response is returned, `RequestLog` is
not completed.

`RetryingClient` waits for some properites of `RequestLog` to be
completed. Therefore, the incomplete `RequestLog` cause `RetryingClient`
to get stuck.

This PR is a temporary workaround to fix the problem. The deadlock
phenomeon will occur easily if a decorator returns a response directly.
I created #5714 that I try to solve this problem ultimately.

Modifications:

- If `getAccessToken()` completes exceptionally, completes log and then
  returns a failed response.

Result:

Fixed a bug where `RetryingClient` gets to deadlock when `OAuth2Client`
fails to obtain an access token.