Remove `RequestLog` dependency in `RetryingClient`
Motivation:

Related: #5708, #4834

If a decorator directly returns a response, the `RequestLog` for the
request won't be complete. Because the `RequestLog` is automatically
completed when it starts to be written to wire by `AbstractRequestHandler`.

In the case of `RetryingClient`, `RequestLog`'s events are used to get
exceptions to apply `RetryRule`. If the `RequestLog` events won't
complete, `RetryingClient` will fall into a deadlock.

Typically, `RequestLog` is automatically completed so it is difficult
for the user to recognize that `RequestLog` must be completed.
Consequently, it is easy to miss.

I propose two changes in this PR to solve the problem.

1. If `RequestLog.requestFirstBytesTransferredTimeNanos` isn't logged
   when a response completes, we can assume that the response is
   returned by a decorator. So a hook is added to a response completion
   event to complete the log manually for that case.

2. `RetryingClient` is refactored not to retry on `RequestLog` but use
   `SplitHttpResponse` and subscribe to headers. An exception when
   headers is received is used to apply `RetryRule`.

Modifications:

- Add a hook that completes `RequestLog` if it is written to wire to
  `ClientUtil`
- Split `handleResponse` into `handleAggregatedResponse` and
  `handleStreamingResponse` in `RetryingClient`
  - Remove `RequestLog` usage in `handleStreamingResponse` and use
    `SplitHttpRequest` to check if an exception is rasied in the
    response.

Result:

- Fixed deadlock that could occur in `RetryingClient` due to `RequestLog`
  not being completed.