Fix a possible deadlock when `OAuth2Client` is used with `RetryingClient` (#5715)
Motivation:

When `OAuth2Client` fails to acquire a granted access token, it directly returns a failed response. As the response is returned, `RequestLog` is not completed.

`RetryingClient` waits for some properties of `RequestLog` to be completed. Therefore, the incomplete `RequestLog` causes `RetryingClient` to get stuck.

This PR is a temporary workaround to fix the problem. The deadlock phenomenon will occur easily if a decorator returns a response directly. I created #5714 in which I will try to solve this problem ultimately.

Modifications:

- If `getAccessToken()` completes exceptionally, completes the log and then returns a failed response.

Result:

Fixed a bug where `RetryingClient` gets to deadlock when `OAuth2Client` fails to obtain an access token.