Support colons in path for server-side (#5676)
Motivation:

Currently, gRPC verb suffix is supported by introducing a `VerbSuffixPathMapping`.
There were several issues with this approach:
- `VerbSuffixPathMapping` was applied inconsistently.
  - For instance, `RouteBuilder path(String pathPattern)` applies `VerbSuffixPathMapping`, but `RouteBuilder path(String prefix, String pathPattern)` doesn't.
- In the context of `ExactPathMapping`, `VerbSuffixPathMapping` was acting as a workaround to support colons in the path. Additionally the support for colons is incomplete since only the last colon is correctly handled.
  - Another side-effect of this is route collisions are incorrectly reported since the verb suffix isn't taken into account.
- Continued support of `VerbSuffixPathMapping` makes supporting colons natively difficult.
- Misc) A path such ending with an escape character failed to start up the server: `/path\\`

The original motivation for #5613 was that the following patterns weren't supported.
- /path/{param}:verb
- /path/literal:verb

I propose that the first case be handled via using `REGEX` `PathMapping` types, and the second case be handled by natively supporting colons in our path parameters.
Note that colons are supported per the [rfc3986](https://datatracker.ietf.org/doc/html/rfc3986#section-3.3), and we already have an issue https://github.com/line/armeria/issues/4577.

1. Supporting `/path/{param}:verb`

I propose that we support this simply by introducing a new `PathMappingType.REGEX`. We can simply check if the last `PathSegment` is a `VerbPathSegment`. If it is a `VerbPathSegment`, then we can just use `PathMappingType.REGEX`.

2. Supporting `/path/literal:verb`

Internally, we represent colons as parameterized variables both in `ParameterizedPathMapping` and `RoutingTrieBuilder`. This makes it difficult to support colons, so I propose that we modify the internal representation to a different character (`\0`). This character isn't a valid path character per [rfc3986](https://datatracker.ietf.org/doc/html/rfc3986#section-3.3), so I believe the impact is minimal.

One side effect of this approach is that `ParameterizedPathMapping#paths` will return null character delimited paths.
e.g. `/v0/:/path` -> `/v0/\0/path`
Having said this, I believe the normal user path doesn't really use this value so it shouldn't matter.

3. Supporting colons in general

When one calls `RouteBuilder#path(String)`, we determine whether to use `ParameterizedPathMapping` or `ExactPathMapping` depending on whether a colon is used.

https://github.com/line/armeria/blob/8ab42847c146b481c72f000dbc1c4d77dc009220/core/src/main/java/com/linecorp/armeria/server/RouteBuilder.java#L546

- If the colon does not start the segment (e.g. `/a:b`), it is trivial to assume that the colon should be matched exactly.
- If the colon does start the segment (e.g. `/:param`), it is ambiguous whether the user intended this to be a literal or parameter.

For this case, I propose the following:
- If a colon is used as-is, `/:param`, then the segment will be used to represent a parameter
- If a colon is escaped, `/\\:param`, then the segment will be treated as a literal

Optimization) `ExactPathMapping` is more performant than `ParameterizedPathMapping` because a simple equality check is done. I propose as an optimization we modify the condition to check if `/:` is contained instead of `:`. As a downside of this approach, the colon escape logic needs to also be added to `ExactPathMapping`. This can be undone if this logic seems too complicated though.

Modifications:

- When `VerbPathSegment` is used, use `RegexPathMapping` for gRPC transcoding.
- Removed `VerbSuffixPathMapping` and related changes in `RoutingTrieBuilder`, `RouteBuilder`, and `ParameterizedPathMapping`
- Modified `RoutingTrieBuilder` and `ParameterizedPathMapping` to use `\0` instead of ':' to represent parameterized path segments. This allows us to use ':' in most `PathMapping` types.
- Relaxed `ParameterizedPathMapping` to act like `ExactPathMapping` when a path segment isn't capturing a path parameter.
- Replace `/\\:` to `/:` in `ParameterizedPathMapping` and `ExactPathMapping` to give users an option to use the first colon as a literal.

Result:

- https://github.com/line/armeria/issues/4577 is closed
- Logic related to gRPC verb suffixes is generalized and cleaned up

<!--
Visit this URL to learn more about how to write a pull request description:
https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->