Support ramping up endpoints for `XdsEndpointGroup` (#5688)
Consider reviewing #5693 before this PR.

Motivation:

This changeset attempts to support [slow start
mode](https://www.envoyproxy.io/docs/envoy/latest/intro/arch_overview/upstream/load_balancing/slow_start).
While simply supporting this is trivial with the existing
`WeightRampingUpStrategy`, ramping up needs to work irregardless of
whether a 1) cluster is updated 2) or the set of endpoints is updated.

The current implementation creates a new `EndpointGroup` whenever one of
the two updates above is triggered.

##### The `ClusterEntry` itself is reconstructed


https://github.com/line/armeria/blob/806556e5a4d27b549c4c6c24f75d96a1447d85d1/xds/src/main/java/com/linecorp/armeria/xds/client/endpoint/ClusterManager.java#L128

##### The set of endpoints is updated


https://github.com/line/armeria/blob/806556e5a4d27b549c4c6c24f75d96a1447d85d1/xds/src/main/java/com/linecorp/armeria/xds/client/endpoint/ClusterEntry.java#L60-L62

Consequentially, endpoints will always be ramped up from the beginning
with the current implementation. e.g. if `endpointA` is added, the new
`EndpointGroup` will trigger all endpoints to be ramped up all together.
As a result `endpointA` won't receive less traffic compared to the other
endpoints.

I propose the following changes to handle this.

##### 1. The `ClusterEntry` itself is reconstructed

I propose that `ClusterEntry` is maintained per cluster name. If a
`ClusterEntry` already exists for a `ClusterSnapshot`, then
`ClusterEntry#updateClusterSnapshot` will be called instead of creating
a new `ClusterEntry`. `SubsetLoadBalancer` also cannot keep
`ClusterSnapshot` as a final field, so the related logic has also been
refactored.

This change is actually probably more in-sync with envoy's
implementation as well.
ref:
https://github.com/envoyproxy/envoy/blob/2e055e40bb93e4bbd97a10b7fedee9c50c5ba8af/source/common/upstream/cluster_manager_impl.cc#L1318

##### 2. The set of endpoints is updated

In order to handle this, I propose that we keep a pool of endpoints in a
class named `DelegatingEndpointGroup`. The `EndpointGroup` will store
the `createdAtNanos` timestamp of previous endpoints, and set these
values if a new `Endpoint` is added. These timestamps will be used to
calculate which ramping up step an `Endpoint` was in previously. Note
that `DelegatingEndpointGroup` is at the outermost layer of
`EndpointGroup`s. This is done to ramp up an `Endpoint` if it was
filtered out by a `HealthCheckedEndpointGroup` and then added back in.

Once a `ClusterSnapshot` is updated,
`DelegatingEndpointGroup#updateClusterSnapshot` will be called.
`DelegatingEndpointGroup` will listen for updates to endpoints, set the
`createdAtNanos` attribute for each endpoint, and call
`ClusterEntry#accept`. It is now possible for `ClusterEntry#accept` to
be called from multiple threads (xDS event loop, health check event
loop, etc..) so I've also added logic to reschedule.

Lastly, in order to fully support xDS's slow-start I've added
`EndpointWeightTransition#aggression`. If a `SlowStartConfig` is
provided, the `EndpointGroup` is constructed with a
`WeightRampingUpStrategy`.

Modifications:

- Modified `ClusterManager` not create a new `ClusterEntry` as long as
the `clusterName` is the same. To support this change,
`ClusterEntry#clusterSnapshotUpdated` is added.
- Added a `DelegatingEndpointGroup` which keeps a reference to the
`EndpointGroup` associated with the `ClusterSnapshot`.
`DelegatingEndpointGroup` also sets `createdAtNanos` to endpoints.
- Added `EndpointWeightTransition#aggression` which allows non-linear
weight transition.
- `WeightRampingUpStrategy` is not set when the `SlowStartConfig`
parameter is set.

Result:

- `XdsEndpointGroup` now supports slow start.

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->

---------

Co-authored-by: minux <minu.song@linecorp.com>