Prevents the connection from being used immediately when `WriteTimeoutException` occurs.
Motivation:

I got a report from LINE internally that the connection was not
terminated but hung when the vm instance was shut down due to
maintainance of the underlying hyperviser.

`operationComplete()` was not called for `Channel.write()` for this
abnormal connection. There was neither a normal response nor a failure
response.
https://github.com/line/armeria/blob/f0ec7cb729d1fb33d238c6ea8fb9af41460ab37d/core/src/main/java/com/linecorp/armeria/client/AbstractHttpRequestHandler.java#L126-L126
As a result, `WriteTimeoutException` occurred and try to reset the
connection.
https://github.com/line/armeria/blob/f0ec7cb729d1fb33d238c6ea8fb9af41460ab37d/core/src/main/java/com/linecorp/armeria/client/AbstractHttpRequestHandler.java#L331-L331

`failAndReset()` calls `Channel.write(Unpooled.EMPTY_BUFFER)` first and
then calls `Channel.close()`. Since `channel.write()` does not respond,
the connection cannot be closed.

Modifications:

- Deactivate `HttpSession` first before proceeding with the reset process.
  - Unhealthy connections will be cleaned up eventually with
    `KeeyAliveHandler` by idle timeout.

Result:

Fixed a bug where a connection was reused after `WriteTimeoutException` occurred.