Fix to propagate `ServiceRequestContext` to `ServerErrorHandler` (#5746)
Motivation:

An event loop of the `Channel` of a request is used to subscribe to a returned `HttpResponse`.
https://github.com/line/armeria/blob/3112b50cb6d2bd0cd31bac17796c5e6a01cc347c/core/src/main/java/com/linecorp/armeria/server/HttpServerHandler.java#L453-L457
The channel event loop is not a context-aware event loop, so `ServiceRequestContext` is not available in the thread local. As a result, request-scoped MDC won't work.

Modifications:

- Push `ServiceRequestContext` before calling `ServerErrorHandler.onServiceException()`

Result:

`ServiceRequestContext` is now correctly propagated to `ServerErrorHandler`.