Change to elegant exception handling
Motivation:
If an unintended file was entered, it threw an exception and terminated. This behavior was frustrating to the user, so changed to a calmer method that simply ignores the file.

Modification:
- Fix(AnnotatedServiceMultipartTest): Change from BAD REQUEST to OK, and change the file name to be more distinguishable.
- Fix(FileAggregatedMultipart): Add filtering
- Fix(Multipart): Add filtering. I wanted to name the filtering method filter,
                  but because the StreamMessage interface already has a filter function,
                  and because the DefaultMultipart class implements both the multipart interface and the StreamMessage interface,
                  I used a different name.

Result:
It no longer raises an acceptance. Instead, unintended files are filtered out before they are written to disk.