`WeightRampingUpStrategy` uses locks to avoid occasional selection failure for `XdsEndpointGroup` (#5799)
Motivation:

https://github.com/line/armeria/issues/5749

While re-investigating this issue, I found that the probably cause is the changes introduced by https://github.com/line/armeria/pull/5693

Previously, when `updateEndpoints` is called the `endpointSelector` was initialized from the same thread 

https://github.com/line/armeria/blob/da7f76a7eebf9b03a51ff68a9245269eee4a5a6d/core/src/main/java/com/linecorp/armeria/client/endpoint/WeightRampingUpStrategy.java#L147-L150

However after the PR above,  `updateEndpoints` doesn't update the `endpointSelector` immediately and instead schedules an update task

https://github.com/line/armeria/blob/191fb7db981d1d47de8ce50be0b0f6d5d9dfa8ee/core/src/main/java/com/linecorp/armeria/client/endpoint/WeightRampingUpStrategy.java#L150

For this reason, even if `updateEndpoints` is called the `endpointSelector` will remain as the `EMPTY_SELECTOR`.

https://github.com/line/armeria/blob/191fb7db981d1d47de8ce50be0b0f6d5d9dfa8ee/core/src/main/java/com/linecorp/armeria/client/endpoint/WeightRampingUpStrategy.java#L128

Normally, this isn't an issue since `refreshEndpoints` will be called after the call to `updateEndpoints` which will in turn complete the `select` call. However, `XdsEndpointGroup` is composed of multiple internal `EndpointGroup`s and only `selectNow` is called. Because the initialization event is never propagated to the `XdsEndpointSelector` occasional failures can occur.

https://github.com/line/armeria/blob/191fb7db981d1d47de8ce50be0b0f6d5d9dfa8ee/xds/src/main/java/com/linecorp/armeria/xds/client/endpoint/DefaultLoadBalancer.java#L61-L84

Modifications:

- Use locks instead of event loops for `updateEndpoints`

Result:

- Endpoint selection calls to `WeightRampingUpStrategy` succeeds before the first call to `updateEndpoints` for  `XdsEndpointGroup`.

<!--
Visit this URL to learn more about how to write a pull request description:
https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->