Fix handling of exceptions in `GrpcExceptionHandlerFunction`. (#5796)
Motivation:
This PR addresses two issues with the `GrpcExceptionHandlerFunction`:
- Peel the exception when calling `apply(RequestContext ctx, Status status, Throwable cause, Metadata metadata)`.
- Convert the cause to a `Status` via `Status.fromThrowable(cause)` before calling the `apply()` method.

The reasoning behind the second change is to resolve the behavior discrepancy between a `StatusRuntimeException` being thrown and a `StatusRuntimeException` being handled with `onError`. When the exception is handled with `onError`, it is converted to a `Status` by the gRPC upstream, and `status.getCause()` is specified when calling the `apply()` method. However, when a `StatusRuntimeException` is thrown, the `StatusRuntimeException` is passed as is.

Modifications:
- Fixed to pass the peeled exception in `UnwrappingGrpcExceptionHandlerFunction`.
- Converted the cause to a `Status` using `Status.fromThrowable(cause)` before calling the `apply()` method.

Result:
- The peeled exception is passed for `GrpcExceptionHandlerFunction.apply(RequestContext ctx, Status status, Throwable cause, Metadata metadata)` method.