Append "all" to "armeria.server.active.requests" in `ServerMetrics`
Motivation:

`MetricCollectingService` uses `.active.requests` to track the number of
the current requests. If `armeria.server` is used as the prefix of the
meter name, the name conflicts with "armeria.server.active.requests" in
`ServerMetrics`.

`ServerMetrics.activeRequests()` counts all requests including
`TransientService` which is excluded by `MetricCollectingService` with
no `TransientServiceOption`. So, semantically, it would be better to add
`.all` to the meter name to indicate that it counts all requests.

Modifications:

- Rename "armeria.server.active.requests" in `ServerMetrics` to
  "armeria.server.active.requests.all".

Result:

The meter names in `ServerMetrics` no longer conflict with
`MetricCollectingService`.