Fix a bug where a protocol violation is not handled by `ServerErrorHandler`
Motivation:

`CorsServerErrorHandler` was added as the first error handler by #5632
`CorsServerErrorHandler` does not handle an exception but delegate it to
the default one or user-defined handler and injects CORS headers.

`CorsServerErrorHandler` overrided `onServiceException` to delegate but
omitted to override `onProtocolViolation`. As a result, a user-defiend
`onProtocolViolation` wasn't invoked.

Modifications:

- Override `onProtocolViolation()` and delegate to `serverErrorHandler`

Result:

A protocol violation is now correctly handled by `ServerErrorHandler.onProtocolViolation()`