Refactor `HaProxy` exception handling (#5805)
Motivation:

While working on https://github.com/line/armeria/pull/5793, I found that
there were unnecessary `ProxyConnectException` related handling in the
request handler. This is most likely due to the fact that
`HaProxyHandler` works differently from the other proxy handlers in that
the session is completed, and then a connection event is published.

This was probably when I was less familiar with netty. I propose that
the `HaProxyHandler` complete the session similarly to the other
handlers. A connection attempt is made first, and then the
`sessionPromise` is succeeded or failed accordingly.

Modifications:

- `HaProxyHandler` doesn't succeed the `sessionPromise` immediately, but
waits for the result of the `connect` future.
- Remove `ProxyConnectException` related logic in all places except for
`HttpSessionHandler#exceptionCaught`

Result:

- The exception handling in `AbstractHttpRequestHandler` and
`HttpSessionHandler` is a little simpler.

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->