Support thrift 0.19 and 0.20 (#5822)
Motivation:

Support for thrift 0.19/0.20 has been requested internally.

Notably, there has been two changes upstream.
- Jakarta EE, Apache HttpClient 5 is now used
[ref](https://github.com/apache/thrift/pull/2746)
- The "true" type is now used when generating Metadata
[ref](https://github.com/apache/thrift/pull/2765/files)

For the first change, test files have been modified so that
HttpClient/Jakarta related classes can be overwritten.

The second change is interesting because generated messages for `typdef`
enums are now correctly set as an `EnumMetadata`. As a result of this,
using `TTextProtocol` with `useNamedEnums=true` now yields the correct
output (`typedef` enums serialized as the enum name). This is consistent
with the other enum serialization results as well.

One downside of this is older version messages (generated with
thrift<0.19) cannot deserialize a typedef enum field by enum name. This
is because a `FieldValueMetaData` is used, and it is not possible to
know which enum the field is linked to unless we parse the thrift file
manually.

Modifications:

- Jakarta EE, Apache HttpClient 5 related dependencies are added
- Extracted `ServletTestUtils` so that
`ThriftOverHttpClientTServletIntegrationTest` can be shared across
modules.
- Extracted `ApacheClientUtils` so that `ThriftDocServiceTest`,
`ThriftOverHttp1Test` can be shared across modules.
  - Modified to use JDK over Apache's `Base64` library
- Generated messages for `typdef` enums are now correctly set, so
`TTextProtocolTest#tTextNamedEnumProtocolReadWriteTest` has been
modified to check if the generated `typedef` value is a enum ordinal or
name.

Result:

- https://github.com/line/armeria/issues/5243

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->