Fix a bug where `INTERNAL_ERROR` is returned when gRPC message length is exceeded (#5824)
Motivation:

If a request or response message exceeds the maximum allowed message
length, `RESOURCE_EXHAUSTED` status should be returned to the client.

In `UnaryServerCall`, an exception raised `ArmeriaMessageFramer` is not
handled by `GrpcExceptionHandlerFunction`. As a result, the exception is
directly passed to the HTTP level `ServerErrorHandler` and a proper gRPC
status is not delivered to the client.

While I was fixing the bug, I found a lot of code that calls
`generateMetadataFromThrowable()` and then `fromThrowable()`. As the
same pattern was used repeatedly, I also refactored
`GrpcExceptionHandlerFunctionUtil` to return `StatusAndMetata` instead.

Modifications:

- Apply `GrpcExceptionHandlerFunction` when an exception is raised in
`UnaryServerCall.doClose()`
- Refactor `GrpcExceptionHandlerFunctionUtil` and its calling site.
- Add `InternalGrpcExceptionHandler` to wrap
`GrpcExceptionHandlerFunction`.
- Remove the incorrect `@Nullable` annotation for
`GrpcExceptionHandlerFunction`.
- Remove `ServerStatusAndMetadata.setResponseContent()` and use
`RESPONSE_CONTENT` property for simpliciy.

Result:

- `GrpcService` now correctly returns `RESOURCE_EXHAUSTED` when a
response message exceeds the maximum allowed length.
- Closes #5818