Fix a leak in `HttpEncodedResponse`
Motivation:

An `HttpData` produced in `HttpEncodedResponse.beforeComplete()` is not
collected by `CollectingSubscriberAndSubscription` and leaked.

```java
  Hint: {10B, pooled, <unknown>}
  com.linecorp.armeria.common.HttpData.wrap(HttpData.java:110)
  com.linecorp.armeria.server.encoding.HttpEncodedResponse.beforeComplete(HttpEncodedResponse.java:163)
  com.linecorp.armeria.common.stream.FilteredStreamMessage.lambda$collect$0(FilteredStreamMessage.java:201)
  java.base/java.util.concurrent.CompletableFuture.uniHandle(CompletableFuture.java:934)
  java.base/java.util.concurrent.CompletableFuture.uniHandleStage(CompletableFuture.java:950)
  java.base/java.util.concurrent.CompletableFuture.handle(CompletableFuture.java:2340)
  com.linecorp.armeria.common.stream.FilteredStreamMessage.collect(FilteredStreamMessage.java:142)
```

`CollectingSubscriberAndSubscription` was designed to only apply
`filter()` of the `upstream.collect()`. I didn't consider that an object
could be published via `onNext()` in `beforeComplete()`.

`CollectingSubscriberAndSubscription` was added to provide an optimized
code path for unary calls. it doesn't seem the code provides a
trival performance improvement but the implemetation is complex and
error-prone.

I was able to fix the code not to leak the data but I didn't want to
additional complexity to it. It might be cleaner to use the Reactive
Streams API instead of keeping the custom `collect()` implemetation.
There will be no change in performance for normal message sizes.

Modifications:

- Remove the custom `collect()` implemtation in `FilteredStreamMessage`.

Result:

Fix a potential leak when sending compressed responses.