Propagate exception from `StreamMessage` to `InputStream`. (#5859)
Motivation:
The `InputStream` returned from `StreamMessage.toInputStream()` should
raise an exception if the original `StreamMessage` is closed with an
exception. Currently, the `InputStream` only returns `-1` in such cases,
which does not inform the caller of the underlying issue.

Modification:
- Updated the implementation to propagate the exception to the
`InputStream` if the original `StreamMessage` is closed with an
exception.

Result:
- The `InputStream` now correctly raises an exception if the original
`StreamMessage` is closed with an exception.